requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2.1
django-channels==0.7.0
django-crispy-forms==2.0
django-multiupload==0.6.1
django-redis==5.2.0
idna==3.4
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
